+++ java/client/src/main/java/org/signal/libsignal/net/ChatConnectionListener.java
@@ -8,6 +8,7 @@ package org.signal.libsignal.net;
 import org.signal.libsignal.internal.FilterExceptions;
 import org.signal.libsignal.internal.Native;
 import org.signal.libsignal.internal.NativeHandleGuard;
+import org.signal.libsignal.internal.TokioAsyncContext;
 
 public interface ChatConnectionListener {
   /**