+++ java/client/src/main/java/org/signal/libsignal/net/FakeChatServer.java
@@ -8,6 +8,7 @@ package org.signal.libsignal.net;
 import org.signal.libsignal.internal.CompletableFuture;
 import org.signal.libsignal.internal.NativeHandleGuard;
 import org.signal.libsignal.internal.NativeTesting;
+import org.signal.libsignal.internal.TokioAsyncContext;
 
 class FakeChatServer extends NativeHandleGuard.SimpleOwner {
   private TokioAsyncContext tokioContext;