+++ rust/bridge/shared/testing/src/convert.rs
@@ -8,6 +8,7 @@ use std::future::Future;
 use futures_util::{AsyncReadExt as _, FutureExt};
 use io::{AsyncInput, InputStream};
 use libsignal_bridge_macros::*;
+use libsignal_bridge_types::net::TokioAsyncContext;
 use libsignal_bridge_types::support::*;
 use libsignal_bridge_types::*;
 use libsignal_protocol::SignalProtocolError;
@@ -71,6 +72,11 @@ async fn TESTING_FutureSuccess(input: u8) -> i32 {
 #[bridge_io(NonSuspendingBackgroundThreadRuntime)]
 async fn TESTING_FutureFailure(_input: u8) -> Result<i32, SignalProtocolError> {
     Err(SignalProtocolError::InvalidArgument("failure".to_string()))
+}
+
+#[bridge_io(TokioAsyncContext)]
+async fn TESTING_TokioAsyncFuture(input: u8) -> i32 {
+    i32::from(input) * 3
 }
 
 #[derive(Clone)]