+++ RELEASE_NOTES.md
@@ -1,2 +1,5 @@
 v0.72.2
 
+- Bridge registration service client to Swift.
+
+- Net/Java: Some CompletableFutures are now cancellable!